`msgtype` field add at  `PostgreSQLLog` and solved data type cast warning and remove `os` deprecated function

+++ src/C++/SSLSocketAcceptor.cpp
@@ -319,7 +319,7 @@ void SSLSocketAcceptor::onConnect( SocketServer& server, socket_handle a, socket
 
   SSL *ssl = SSL_new(m_ctx);
   SSL_clear(ssl);
-  BIO *sBio = BIO_new_socket(s, BIO_NOCLOSE); //Unfortunately OpenSSL assumes socket is int
+  BIO *sBio = BIO_new_socket(static_cast<int>( s ), BIO_NOCLOSE ); //Unfortunately OpenSSL assumes socket is int
   SSL_set_bio(ssl, sBio, sBio);
   // TODO - check this
   SSL_set_app_data(ssl, m_revocationStore);
@@ -412,7 +412,7 @@ int SSLSocketAcceptor::passwordHandleCallback(char *buf, size_t bufsize,
     return -1;
 
   std::strcpy(buf, m_password.c_str());
-  return m_password.length();
+  return static_cast<int>( m_password.length( ) );
 }
 }
